[BOOKINGSG-9335][IT] Refactor files and add styling
