+++ src/time-slot-bar-week/consts.ts
@@ -1,0 +1,3 @@
+export const SLOT_HEIGHT = 12; // px
+export const SLOT_GAP = 4; // px
+export const ROW_HEIGHT = SLOT_HEIGHT + SLOT_GAP;
